chore: move traits to register files
