+++ src/registers/mod.rs
@@ -14,10 +14,7 @@ pub mod ctrl8_xl;
 pub mod ctrl9_xl;
 pub mod drd_src;
 pub mod emb_func_status;
-pub mod fifo_ctrl;
-pub mod fifo_data_out;
-pub mod fifo_data_out_tag;
-pub mod fifo_status;
+pub mod fifo;
 pub mod free_fall;
 pub mod fsm_status;
 pub mod func_cfg_access;
@@ -27,7 +24,9 @@ pub mod internal_freq_fine;
 pub mod md_cfg;
 pub mod mlc_status;
 pub mod ofs_usr;
+pub mod out_g;
 pub mod out_temp;
+pub mod out_xl;
 pub mod pin_ctrl;
 pub mod status_master;
 pub mod status_reg;
@@ -53,10 +52,7 @@ pub use ctrl9_xl::*;
 pub use ctrl10_c::*;
 pub use drd_src::*;
 pub use emb_func_status::*;
-pub use fifo_ctrl::*;
-pub use fifo_data_out::*;
-pub use fifo_data_out_tag::*;
-pub use fifo_status::*;
+pub use fifo::*;
 pub use free_fall::*;
 pub use fsm_status::*;
 pub use func_cfg_access::*;
@@ -66,7 +62,9 @@ pub use internal_freq_fine::*;
 pub use md_cfg::*;
 pub use mlc_status::*;
 pub use ofs_usr::*;
+pub use out_g::*;
 pub use out_temp::*;
+pub use out_xl::*;
 pub use pin_ctrl::*;
 pub use status_master::*;
 pub use status_reg::*;
